Email util added for automation
